Add files via upload
The Caesar Cipher can encrypt and decrypt a string based on a provided key.  The Caesar Cracker can crack a string encoded with a Caesar Cypher and provide the key used to encrypt the message.  Different languages can be decoded by adjusting the mostCommon variable which is associated with the most common letter in the given language.  Vigenere Cipher and Breaker builds upon the Caesar Cypher by essentially utilizing multiple Caesar Ciphers within one message, making the encrypted message more secure.  The Vigenere Breaker Multiple Languages program can crack any Vigenere encrypted message in any language, provided the list of dictionaries for each language.  In addition to decoding the message, it can also return the original language the string was decoded to.